move navigation bar to different module
