+++ src/components/NavigationBar/LocaleSwitcher.js
