+++ src/components/NavigationBar/LocaleSwitcher.module.css
